Tweak the custom parameter handling.
Custom parameters are updated by user input to the UI.  When the
user updates these, we want to review what the user has entered to
determine if we are doing local access/cloud access or both.

Also verify we have all the parameters set for the type(s) of
access it looks like they want.

At this point, we can then do our initial connections/queries and
determine what nodes need to be created.