Make nodes static
There isn't much variation in the nodes so with a bit of
reorganization it is possible to use statically defined nodes.

This simplifies the code base by making all the code that
builds the profile files unnecessary and also simplifies the
node updates.  No reason to build a mapping of what variables
are assigned to what node, just define these statically as well.

As part of this, we split the status node into three nodes:

1  - the main status node (common for all setups)
2  - the solar status node for systems with solar
3  - the generator status node for systems with a generator

The solar and generator status nodes are (or can be) conditionally
created.

There should be no difference in the setup node or the main controller
node.  However, it might make sense to combine the main controller
node and the status node into a single node as a future change.